Bump `windows-core` range to `0.56-0.61` and regenerate with `0.59`
`windows-bindgen 0.59` now generates way less code since it no longer
emits functions for which types are not enabled - that we don't use
anyway!  And the added bonus is that this generated code is compatible
with `windows-core` all the way back to release `0.56`, meaning we don't
force everyone to immediately upgrade once they're already on `0.56`.
