+++ .github/workflows/rust.yml
@@ -25,12 +25,15 @@ jobs:
           - nightly
         versions:
           - ""
-          - "-Zminimal-versions"
+          - -Zminimal-versions
         exclude:
           # package `regex v1.11.1` cannot be built because it requires rustc
           # 1.65 or newer, while the currently active rustc version is 1.62.0
           - toolchain: "1.62"
             versions: ""
+          # Older proc-macro2 crates fail to compile on the latest nightly
+          - toolchain: nightly
+            versions: -Zminimal-versions
     runs-on: ${{ matrix.runs-on }}
     steps:
       - name: Checkout
@@ -54,13 +57,13 @@ jobs:
       matrix:
         versions:
           - ""
-          - "-Zminimal-versions"
+          - -Zminimal-versions
         toolchain:
           - stable
           - nightly
         include:
           # Without `-Zminimal-versions` a too recent bumpalo version is selected. Newer versions use `edition = "2021"`.
-          - versions: "-Zminimal-versions"
+          - versions: -Zminimal-versions
           - toolchain: "1.62"
     runs-on: ubuntu-latest
     steps:
@@ -119,7 +122,7 @@ jobs:
           # - sparcv9-sun-solaris
         versions:
           - ""
-          - "-Zminimal-versions"
+          - -Zminimal-versions
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
@@ -192,7 +195,7 @@ jobs:
           - x86_64-apple-ios
         versions:
           - ""
-          - "-Zminimal-versions"
+          - -Zminimal-versions
         exclude:
           # package `regex v1.11.1` cannot be built because it requires rustc
           # 1.65 or newer, while the currently active rustc version is 1.62.0
@@ -226,7 +229,7 @@ jobs:
           - nightly
         versions:
           - ""
-          - "-Zminimal-versions"
+          - -Zminimal-versions
         exclude:
           # package `cc v1.2.17` cannot be built because it requires rustc 1.63
           # or newer, while the currently active rustc version is 1.62.0
