+++ api_gen/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 publish = false
 
 [dependencies]
-windows-bindgen = "0.56"
+windows-bindgen = "0.59"
 
 # Dissociate this crate from the root workspace to allow older Rust to build
 # the main crate within the workspace, without observing higher `rust-version`
